Add dynamic-values sub-command
